Revert "rtpmux: fix PROP_TIMESTAMP_OFFSET range problems"
This broke API, so we need a better solution!

This reverts commit c7579d31a6e9d788e94b83258309063d0aae481e.

+++ gst/rtpmanager/gstrtpmux.h
@@ -43,7 +43,7 @@ typedef struct _GstRTPMuxClass GstRTPMuxClass;
 typedef struct
 {
   gboolean have_timestamp_offset;
-  guint32 timestamp_offset;
+  guint timestamp_offset;
 
   GstSegment segment;
 
@@ -66,9 +66,8 @@ struct _GstRTPMux
   guint32 ts_base;
   guint16 seqnum_base;
 
-  gint64 ts_offset;
-  gint seqnum_offset;
-
+  gint32 ts_offset;
+  gint16 seqnum_offset;
   guint16 seqnum;               /* protected by object lock */
   guint ssrc;
   guint current_ssrc;
